Improves to multiple tools - see message for more info
BinaryInterpreterWPF: Add State. This is not yet fully supported, still trying to find the exact pattern. It has some sort of unrealscript in it, just need to figure out how to slice it out and parse.

Bio2DAEditorWPF: Do not parse Default__ objects

PackageEditorWPF: Optimize performance by not reloading classlist every time you click an object, instead caching the list. WE will need to write a hook to refresh this list when headers change or an import or export is added. Also added a welcome tab that shows when 0 or nothing is selected

InterpreterWPF: Hide editor set when the export is unloaded. Optimize array properties to generate using the same method as other properties. Change some code over to pattern code style. Split ByteProperty and EnumProperty out and mark ByteProperties that are actually enums. Support EnumProperty, StringRefPRoperty value setting in the editorset. StrRefProperty will preview the value in editorset, like objectproperty now - needs a way to improve multiline display in the upropertytreeviewitem.

Added some comments for possible future optimizations in a few files from waht I could see in my performance profiling sessions.