Multiple bugfixes for WPF tools
BinaryInterpreterWPF: Class scan now can handle some stuff when things are imports, when we previously did not expect them to be. Output of Class Scan will be shown up to the point of an error if one occurs.

InterpreterWPF: Change how ArrayProperty text is displayed

ScriptEditorWPF: Can now byteshift bytecode parsing for state objects. This part of UI is not done yet

PropertyCollection: Writing Struct Properties that are Immutable will check if the final property is a NoneProperty. If it is, it will not append one (this is due to InterpreterWPF showing NoneProperty at end of Struct)