Sequence Editor Experiments (#336)
Adds the following experiments:

Add director and camera Interp templates for controlling the dialogue wheel.
Update missing variable links of selected or all Interps in the selected sequence. It adds any new named InterpGroup to the links, removes any link that no longer matches an InterpGroup of the InterpData, and keeps anyone that hasn't been touched. It does not change the Data or Anchor links, as those are added by default.

Commits:

* SeqEd: Add experiments to add missing var links, and templates for dialogue wheel control

* SeqEd: Update experiment to add missing variable links to fully update them.

* SeqEd: Add default InterpLength to dialogue wheel experiments

* SeqEd: Add experiments to update the variable links of all the interps that are selected or in a sequence

* SeqEd: Remove singular interp experiment from the menu, calling the selected interps directly is enough. Add custom min and max vars prop to new var links.

* SeqEd: Change checks in UpdateVarLinks to happen regardless of inLoop. Fix LinkDesc typo.

* SeqEd: Fix style of new experiments