Experiments to clean conversations (#351)
The main experiment allows for stripping a BioConversation of everything that is not strictly needed for it to work, allowing you to not bring unnecessary stuff when cloning them to other files. It also renames everything related to it with the goal of allowing you to use it even if the original was loaded into memory, though for the moment that isn't fixing it, so right now it's just an aesthethics thing.
The other experiments are the steps required for the main experiment abstracted so they can be used independently.